Resolve "Convert all count* Functions to Return longs"

Closes #393

See merge request echothree/echothree!432

+++ src/java/com/echothree/control/user/search/common/result/GetContactMechanismResultsResult.java
@@ -23,8 +23,8 @@ import java.util.List;
 public interface GetContactMechanismResultsResult
         extends BaseResult {
     
-    Integer getContactMechanismResultCount();
-    void setContactMechanismResultCount(Integer contactMechanismResultCount);
+    Long getContactMechanismResultCount();
+    void setContactMechanismResultCount(Long contactMechanismResultCount);
 
     List<ContactMechanismResultTransfer> getContactMechanismResults();
     void setContactMechanismResults(List<ContactMechanismResultTransfer> contactMechanismResults);
