+++ src/java/com/echothree/control/user/search/common/result/GetItemResultsResult.java
@@ -23,8 +23,8 @@ import java.util.List;
 public interface GetItemResultsResult
         extends BaseResult {
     
-    Integer getItemResultCount();
-    void setItemResultCount(Integer itemResultCount);
+    Long getItemResultCount();
+    void setItemResultCount(Long itemResultCount);
 
     List<ItemResultTransfer> getItemResults();
     void setItemResults(List<ItemResultTransfer> itemResults);
